+++ src/reducers/books.js
@@ -3,14 +3,11 @@ import { CREATE_BOOK, REMOVE_BOOK } from '../actions/index';
 const booksReducer = (state = [], action) => {
   switch (action.type) {
     case CREATE_BOOK:
-      return { books: [...state.books, action.book] };
+      return [...state, action.book];
     case REMOVE_BOOK:
-      return {
-        books:
-        [...state.books.slice(0, action.book),
-          ...state.books.slice(action.book + 1,
-            state.books.length)],
-      };
+      return [...state.slice(0, action.book),
+        ...state.slice(action.book + 1,
+          state.length)];
     default:
       return state;
   }