+++ src/reducers/index.js
@@ -1,5 +1,7 @@
 import { combineReducers } from 'redux';
 import booksReducer from './books';
-import changeFilter from './filter'
+import changeFilter from './filter';
 
-export default combineReducers({ books: booksReducer, filter: changeFilter });
+const rootReducer = combineReducers({ books: booksReducer, filter: changeFilter });
+
+export default rootReducer;