+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Books List and Form
 - Books Component
 - Correct handleChange on booksForm component
+- Add filter functionality
 
 ### Fixed 
 - Moved Books List and Book Form to containers